+++ src/main/java/org/example/delivery/common/JwtUtil.java
@@ -1,4 +1,4 @@
-package org.example.delivery.config;
+package org.example.delivery.common;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.JwtException;
@@ -9,7 +9,7 @@ import jakarta.annotation.PostConstruct;
 import org.example.delivery.domain.user.enums.UserRole;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import org.springframework.util.StringUtils;
+
 import java.security.Key;
 import java.util.Date;
