--- src/main/java/org/example/delivery/domain/common/exception/CustomResponse.java
@@ -1,4 +1,0 @@
-package org.example.delivery.domain.common.exception;
-
-//public class CustomResponse implements org.springframework.web.ErrorResponse {
-//}
